Backend routes

+++ app/api/board_routes.py
@@ -143,9 +143,10 @@ def delete_boards(boardId):
         # abort(404, {"message": "Board not Found"})
         return {"message": "Board not Found"}
     
-    # Commit the deletion to the db
-    db.session.delete(user_board)
-    db.session.commit()
+    if user_board.user_id == current_user.id:
+        # Commit the deletion to the db
+        db.session.delete(user_board)
+        db.session.commit()
 
     # Return a successful deletion response
     return jsonify({"message": "Board has been Deleted successfully" }), 200
